[RENAME] #1467 CommentSection -> CommentInputContainer
