+++ src/feature/comment/component/index.js
@@ -5,7 +5,7 @@ export { default as CommentList } from './CommentList/CommentList';
 export { default as CommentListErrorFallback } from './CommentList/CommentListErrorFallback';
 export { default as CommentListSuspense } from './CommentList/CommentListSuspense';
 
-export { default as CommentSection } from './CommentSection/CommentSection';
+export { default as CommentInputContainer } from './CommentInputContainer/CommentInputContainer';
 
 export { default as CommentInput } from './CommentInput/CommentInput';
